requirements.txt
boto3
moto
Pillow
